win32 cmake configueria

+++ libraries/repro-mysql/src/mysql-async.cpp
@@ -102,7 +102,7 @@ MYSQL_FIELD* statement_async::field( size_t i ) const
 	if ( i >= column_count_ )
 		throw repro::Ex("invalid field!");
 
-	return mysql_fetch_field_direct(prepare_meta_result_.get(), i);
+	return mysql_fetch_field_direct(prepare_meta_result_.get(), (unsigned long) i);
 }
 
 size_t statement_async::param_count() const {
